Move register label to model
