+++ src/components/model/RegisterLabel.vue
